try to fix error

+++ FVCalc.java
@@ -6,6 +6,6 @@ public class FVCalc {
 		int n = Integer.parseInt(args[2]);
 		double rate1 = rate/100 + 1;
 		double futureValue = currentValue * Math.pow (rate1 ,2);
-		System.out.println(" After " + n + " years, a $" + currentValue + " saved at " + rate + "% will yield $" + (int) futureValue );
+		System.out.println(" After " + n + " years, a $" + (int) currentValue + " saved at " + rate + "% will yield $" + (int) futureValue );
 	}
 }
